requirements.txt
pygame>=2.1
django>=4
djangorestframework>=3.13
pygame_gui>=0.6
pymemcache>=3.5
requests>=2.28
django-mathfilters>=1.0